pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bss-forecast"
version = "1.0.0"
description = "Blue Summit Supplies sales analytics and forecasting"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "scikit-learn>=1.3",
    "prophet>=1.1",
    "pmdarima>=2.0",
    "mlxtend>=0.23",
    "rapidfuzz>=3.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "statsmodels>=0.14",
]

[tool.setuptools.packages.find]
include = ["bss*"]
